+++ src/middleware.ts
@@ -168,7 +168,7 @@ export async function middleware(request: NextRequest) {
 export const config = {
   matcher: [
     {
-      source: "/((?!_next/static|_next/image|favicon.ico|public/).*)",
+      source: "/((?!_next/static|_next/image|favicon.ico|monaco-workers|public/).*)",
       missing: [
         { type: "header", key: "next-router-prefetch" },
         { type: "header", key: "purpose", value: "prefetch" },